ParamStore - Separated impl from contract into separate files
